requirements.txt
langchain==0.3.17
langchain-openai==0.2.14
python-dotenv==1.0.1
openai==1.59.8
